correct logic in package repository: - create new modules and split all repositories in this modules
