+++ app/service/repositories/__init__.py
@@ -1,14 +1,13 @@
-from .repositories import (
-    OrganizerRepository,
-    ContactRepository,
-    UserRepository,
-    UserCompanyRepository,
-    LinkCodeRepository,
-    ProductRepository,
-    ProductVersionRepository,
-    SupplyRepository,
-    SupplyProductRepository,
-    ExpenseCompanyRepository,
-    ExpenseSupplierRepository,
-    StatisticRepository
-)
+from .organizer import OrganizerRepository
+from .contract import ContactRepository
+from .user import UserRepository
+from .user_company import UserCompanyRepository
+from .link_code import LinkCodeRepository
+from .product import ProductRepository
+from .product_version import ProductVersionRepository
+from .supply import SupplyRepository
+from .supply_product import SupplyProductRepository
+from .expense_company import ExpenseCompanyRepository
+from .expense_supplier import ExpenseSupplierRepository
+
+from .statistic import StatisticRepository
